Renames class files and update reference in other files
