+++ app/routes/User.js
@@ -2,8 +2,8 @@
  * Created by cyrielo on 11/8/16.
  */
 import express from 'express';
-import Authenticate from './../middleware/authenticate';
-import Authorize from './../middleware/authorize';
+import Authenticate from '../middleware/Authenticate';
+import Authorize from '../middleware/Authorize';
 import UserCtrl from '../controllers/Users';
 
 /**
