+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "istanbul": 
     "./node_modules/.bin/babel-node ./node_modules/babel-istanbul/lib/cli.js cover node_modules/.bin/_mocha --reporter lcovonly -- -- test/*",
-    "seed": "NODE_ENV=test ./node_modules/.bin/babel-node app/seeders/seed.js",
+    "seed": "NODE_ENV=test ./node_modules/.bin/babel-node app/seeders/Seeder.js",
     "db-setup": "dropdb --if-exists test_doc_manager && createdb test_doc_manager",
     "test": "npm run db-setup && npm run seed && NODE_ENV=test npm run istanbul",
     "start": "npm run db-setup && npm run seed && NODE_ENV=test nodemon ./node_modules/.bin/babel-node server.js",
